requirements.txt
boto==2.49.0
boto3==1.9.20
botocore==1.12.20
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.6
gensim==3.6.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mysql-connector-python==8.0.12
numpy==1.15.2
protobuf==3.6.1
python-dateutil==2.7.3
requests==2.19.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
urllib3==1.23
Werkzeug==0.14.1

